이미지가 포함된 리뷰 등록 구현, 엔티디로 값 받는 부분 실패

+++ tripleCMS/tripleCMS/src/main/java/com/tripleCMS/tripleCMS/model/History.java
@@ -21,7 +21,7 @@ public class History {
 
     // 이력 번호
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int historyNo;
+    private long historyNo;
 
     // 이벤트 타입
     @Enumerated(EnumType.STRING)
